[net] make SockAddr::HashCode() return size_t
This patch changes the return type of SockAddr::HashCode() from uint32_t
to size_t (since the underlying hash function effectively operates with
64-bit integers on x86_64 anyway) and brings more consistency into
the specialization of the std::hash<T>::operator()(const T&) for the
Sockaddr type.  This addresses UBSAN warnings on implicit conversion of
integers (when enabled) and makes SockAddr::HashCode() consistent with
the expectations for std::hash<T>::operator() template specialization.

Reviewed-on: http://gerrit.cloudera.org:8080/23876
Tested-by: Alexey Serbin <alexey@apache.org>

+++ src/kudu/util/net/sockaddr.cc
@@ -290,7 +290,7 @@ void Sockaddr::set_length(socklen_t len) {
                             sizeof(storage_) - len_);
 }
 
-uint32_t Sockaddr::HashCode() const {
+size_t Sockaddr::HashCode() const {
   return HashStringThoroughly(reinterpret_cast<const char*>(&storage_), len_);
 }
